Rename package to match directory structure
