+++ catalogue_pipeline/matcher/src/main/scala/uk/ac/wellcome/platform/matcher/locking/DynamoRowLockDao.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.matcher.lockable
+package uk.ac.wellcome.platform.matcher.locking
 
 import java.time.{Duration, Instant}
